Add helm chart demonstrating multi model serving using k8s (#887)
This PR adds an example of a Kubernetes helm chart that is used to
deploy the vLLM server with multiple models using a single Gaudi card.
The templates in the helm chart follow the same type of deployment
described in the [vLLM Kubernetes
documentation](https://docs.vllm.ai/en/latest/deployment/k8s.html) where
it uses a secret for the HF token, PVC for the HF hub cache, and a
deployment/service for the vLLM server. The helm chart `values.yaml`
file deploys the vLLM server using `mistralai/Mistral-7B-Instruct-v0.3`
and `meta-llama/Llama-3.1-8B-Instruct` by default. These can be modified
as needed.

Signed-off-by: Dina Suehiro Jones <dina.s.jones@intel.com>