optimize looping over attention matrix and avoiding fully padded chunks
A large padded attention matrix is created with bucketization of q and k dimensions where the relevant part is at the bottom right.
However, we don't need to gather tokens and calculate the matmul for full chunks.
Modifying the boundries of the for loops allow us to loop only over the causal meaningful chunks of the matrix.