optimize gather_kv_tokens to gather only effective batch
For some chunks only part of the samples in the batch are required and the rest are padded.
Similarly to the optimization controlled by DO_AVOID_FULLY_PADDED_CHUNKS which minimizes the batch before gathers the q chunk,
the added optimization is controlled by DO_REDUCE_KV_GATHER_BATCH and reduces the batch before gathring kv tokens.
This optimization is based on reducing the batch and bucketizing, forcing us to break the graph.