[SW-224431] Fix fp8 measurement for mixtral (#1119)
Previously it was only checking if it is using quant_config and choosing
VllmMixtureOfExpertsOpFP8 as OP, which only difference is that when
measuring scales it is assuming block quant. This will only happen when
we are using Fp8MoEMethod as quant_method.
Kwargs in moe_op call had to be disabled, beacuse of different apis of
FP8 and unquantized

---------

Signed-off-by: kwisniewski98 <kwisniewski@habana.ai>

Signed-off-by: Chendi Xue <chendi.xue@intel.com>