add torch profiler for the LLM engine (#1114)
This PR enabled a profiler for each step of the LLMEngine. The following
4 ENVs are used to control the profiler:

- VLLM_ENGINE_PROFILER_ENABLED, set to true to enable device profiler.
- VLLM_ENGINE_PROFILER_WARMUP_STEPS, number of steps to ignore for
profiling.
- VLLM_ENGINE_PROFILER_STEPS, number of steps to capture for profiling.
- VLLM_ENGINE_PROFILER_REPEAT, number of cycles for (warmup + profile).

> Please refer to
[torch.profiler.schedule](https://pytorch.org/docs/stable/profiler.html#torch.profiler.schedule)
for more details about the profiler schedule arguments.

> The step in profiling means a step of the LLM engine and **exclude**
the profile and warmup run in HabanaModelRunner.

Please use `export VLLM_PROFILER_ENABLED=True` to enable the high-level
vLLM profiler and use the result to choose the steps for detailed
LLMEngine profiling.