Pad the batch size to a bucket batch size
It might result in runtime error if the batch size is not padded to a
bucket batch size.

E.g.

[rank0]:   File "/workspace/aicse.vllm-habana/vllm/model_executor/custom_op.py", line 25, in forward
[rank0]:     return self._forward_method(*args, **kwargs)
[rank0]:   File "/workspace/aicse.vllm-habana/vllm/model_executor/layers/rotary_embedding.py", line 268, in forward_hpu
[rank0]:     key = key.view(num_tokens, -1, self.head_size)
[rank0]: RuntimeError: shape '[9216, -1, 128]' is invalid for input of size 4128768

Note before commit 0f513bd, the batch size passed to
self.warmup_scenario() was the maximum batch size in all buckets for
prompt, so requring the batch size passed to self.warup_scenario()
matches one of a bucket after commit 0f513bd is logical.

Signed-off-by: Haihao Xiang <haihao.xiang@intel.com>