Improve the compatibility and robustness of the device name config (#1726)
## Purpose
The recent improvement to the device name configure to support both
string and list is not complete and not compatible with the original
config.
We will support both string and list in a compatible way:
1. if string is specified, the string will be used as the device name.
2. if list is specified, each rank will use the corresponding device
name in the list. Raise an error if the number of ranks is larger than
the number of device names specified.