First token from decode for non stream case (#1745)
## Purpose
For request with stream=False, the output will be incorrect if we use
the first chunk from prefill and skipped the first trunk from decode.
Because when stream=False, each decode chunk possibly contains multiple
tokens which result wrong output if skipped.

We should not use prefill output and skipped the first decode chunk for
non-steam case.