[DeepSeek R1] Optimize the serial fetching with responsive fetching (#1844)
## Purpose
The current async fetching is done in the following way:
get request -> wait key -> fetch data -> get request -> wait key ->
fetch data -> ...
This poses a serious problem when there is a key is unavailable and the
whole process will wait on the key until timeout happens. During the
wait period, no more new quest queued can be processed and will not
response to these new request even the whole process is idle but wait
for the current key to be available.

We need to improve the serial fetching to a more responsive fetching. 

This PR implement a solution that we will wait on a list of keys and any
key available will be dispatched to fetching. And during the wait
period, we will also make sure the new requests for fetching and join to
the wait list so that the whole process will not be blocked by previous
key(s) is not available.

We also optimized the kv cache key use a signal key to mitigate the
troubling GC issue.