auto config parameters and ENVs for INC (#1842)
INC is enabled by setting `QUANT_CONFIG`, and the relative parameters
and ENVs are configured automatically:
1. Set `PT_HPU_WEIGHT_SHARING=0` if `QUANT_CONFIG` is set.
2. For a model with quantized weights: set `quantization=None` and vLLM
will set it to the `quantization_config.quant_method` from the
`config.json` of the model.
3. For a model with bf16/fp16 weights: set `quantization='inc'`.
3. Set `kv_cache_dtype='fp8_inc` if running in `quantize` mode and the
KVCache is not blocked in the `blocklist` of `QUANT_CONFIG`, otherwise
set `kv_cache_dtype='auto`.
4. Set `weights_load_device='cpu'` and `VLLM_HPU_CONVERT_TO_FP8UZ=true`
on Gaudi2 to enable `fp8_e4m3fn->fp8_e4m3fnuz` conversion for models
with quantized weights.
5. Set `VLLM_DISABLE_MARK_SCALES_AS_CONST=true` by default.

---------

Signed-off-by: Youlei Yang <youlei.yang@intel.com>
Co-authored-by: Liu, Yi4 <yi4.liu@intel.com>