move all the tensors in the module to HPU (#2025)
`--weights-load-device cpu` is set by default on Gaudi2 for the
DeepSeek-R1 model to enable the conversion from `fp8_e4m3fn` to
`fp8_e4m3fnuz`. While the `W_UV` and `W_UK_T` cannot move to HPU by
`model.to("hpu")`, which causes extra memory consumption and results in
OOM. This PR moves all the Tensors, List/Tuple of Tensors and Dictionary
with Tensor values to HPU to fix this issue.

Signed-off-by: Youlei Yang <youlei.yang@intel.com>