[DeepSeek R1] P/D chunked prefill warmup with chunk size using context blocks (#2139)
Implement warmup for P/D chunked prefill when prefill chunk size is
specified. This is a simpler warmup case for chunked prefill. When chunk
size is not specified, the number of warmup cases will be increased by
times of(max_seq_len / block_size) which is not realistic to implement.

This implement using the max_model_len and the chunk size to calculate
the possible number of context chunks and warmup with this context
chunks.